pyproject.toml
[project]
name = "epubhv"
version = "0.5.0"
description = "make your epub books vertical or horizontal."
authors = [{ name = "yihong0618", email = "zouzou0208@gmail.com" }]

dependencies = [
    "beautifulsoup4==4.12.2",
    "cssutils==2.7.1",
    "lxml==4.9.3",
    "opencc-python-reimplemented==0.1.7",
    "soupsieve==2.5",
    "typed-argument-parser==1.8.1",
    "jieba",
    "pypinyin",
    "unidic_lite",
    "fugashi",
    "langdetect",
    "ToJyutping",
]

requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/yihong0618/epubhv"

[project.scripts]
epubhv = "epubhv.cli:main"
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
include = ["epubhv"]

[tool.pdm.scripts]
format = "black ."
lint = "black . --check"
typing = "pyright"
test = "pytest -vv"
all = { composite = ["lint", "typing", "test"] }

[tool.pdm.dev-dependencies]
test = ["pytest>=7.4.2"]
style = ["black>=23.9.1"]
typing = ["pyright>=1.1.327", "types-beautifulsoup4>=4.12.0.6"]


[tool.pyright]
include = ["epubhv", "tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs",
    "epubhv/yomituki.py"
]
defineConstant = { DEBUG = true }

reportMissingImports = true
reportMissingTypeStubs = false
